Begin mit der CLI Entwicklung / Schritt: Installation

+++ lib/constlib/__init__.py
@@ -1,6 +1,7 @@
 import socket
 import asyncio
 import os
+import sys
 import aiofiles
 import aiofiles.os
 
@@ -18,18 +19,21 @@ class Const(BaseObj):
             self.hostname = ''
         self.pid = os.getpid()
         self.reload = False
+        self.app = sys.argv[0].split('/')[-1].split('.')[0]
         self.is_docker = False
         
     async def _ainit(self):
         self.loop = asyncio.get_event_loop()
-        try:
-            async with aiofiles.open(PIDFILE, 'w') as f:
-                await f.write(str(self.pid))
-        except Exception as e:
-            print(e)
+        if self.app == "supervisord":
+            try:
+                async with aiofiles.open(PIDFILE, 'w') as f:
+                    await f.write(str(self.pid))
+            except Exception as e:
+                print(e)
 
     async def _astop(self):
-        await aiofiles.os.remove(PIDFILE)
+        if self.app == "supervisord":
+            await aiofiles.os.remove(PIDFILE)
 
         
 '''
